Transparent osmosis-swap e2e test working

+++ crates/tests/src/e2e/helpers.rs
@@ -668,7 +668,7 @@ fn make_hermes_chain_config_for_cosmos(
         Value::String(key_dir.to_string_lossy().to_string()),
     );
     chain.insert("store_prefix".to_owned(), Value::String("ibc".to_owned()));
-    chain.insert("max_gas".to_owned(), Value::Integer(500_000));
+    chain.insert("max_gas".to_owned(), Value::Integer(500_000_000));
     chain.insert("gas_multiplier".to_owned(), Value::Float(2.3));
     let mut table = toml::map::Map::new();
     if let CosmosChainType::Osmosis = chain_type {
